change type id from bigint to int in scripts create tables

+++ src/main/resources/create_table_groups.sql
@@ -4,7 +4,7 @@
 
 CREATE TABLE public.groups
 (
-    group_id bigint NOT NULL DEFAULT nextval('groups_group_id_seq'::regclass),
+    group_id integer NOT NULL DEFAULT nextval('groups_group_id_seq'::regclass),
     group_name character varying(10) COLLATE pg_catalog."default" NOT NULL,
     CONSTRAINT groups_pkey PRIMARY KEY (group_id)
 )
